requirements.txt
aiofiles==23.2.1
SpeechRecognition==3.14.1
sounddevice==0.5.1
soundfile==0.13.1
google-generativeai==0.8.5
groq==0.25.0
ffmpeg-python==0.2.0
fastapi==0.115.11
openai==1.66.5
pandas==2.2.3
nltk==3.9.1
noisereduce==3.0.3
colorlog==6.9.0
matplotlib==3.9.4
tiktoken==0.9.0
pyttsx3==2.98
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
pytest==8.3.5
PyMuPDF==1.25.4
pyparsing==3.2.1
PyPDF2==3.0.1
pygame==2.6.1
pydub==0.25.1
pydantic==2.10.6
PyAudio==0.2.14
pyarrow==19.0.1
pillow==10.4.0
nest-asyncio==1.6.0
google-genai==1.15.0
Flask==3.1.0
gTTS